requirements.txt
selenium>=4.15.0
pandas>=2.0.0
openpyxl>=3.1.0
webdriver-manager>=4.0.0
flask>=2.3.0
gunicorn>=21.2.0
python-dotenv>=1.0.0

